pyproject.toml
[project]
name = "agents021"
version = "0.1.0"
description = "A modern multi-agent system featuring dynamic synthesis and evolutionary optimization."
authors = [
    { name = "Xia Ruilong", email = "ruilongxia@gmail.com" }
]
dependencies = [
    "agno>=2.5.17",
    "datasets>=4.8.4",
    "ddgs>=9.14.0",
    "fastapi>=0.136.0",
    "httpx[socks]>=0.28.1",
    "lancedb>=0.30.2",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "prompt-toolkit>=3.0.52",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.2.2",
    "rich>=13.7.1",
    "sqlalchemy>=2.0.49",
    "typer>=0.24.1",
    "uvicorn>=0.44.0",
]
readme = "README.md"
requires-python = ">=3.13"

[project.scripts]
trinity = "core.cli.main:app"

[project.optional-dependencies]
dev = [
    "mypy>=1.9.0",
    "pytest>=8.1.1",
    "ruff>=0.3.4",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests", "core/tests"]
python_files = ["test_*.py"]
pythonpath = "."

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["core"]
